move types.tsx to parent folder
